refactor: windows mdm code restructure
The MDM code had 3 files each with the same name for each part of the protocol, this became confusing as my code editor only shows the files name in the toolbar. This restructure fixes that up and also abstracts code that is repeated throughout the handlers into a common package.