refactor: redone services to allows for MDM enrollmement
The settings, certificates and devices services/store were redone to allow for device enrollment. They also fix possible errors caused by multiple Go routines accessing and changing the settings at the same time by using a Mutex. Services and stores were split to make it easier for databases to be interchanged in the future.
Note: WinMDM Enrollment is broken in this commit.