feat: added datastore abstraction + basic version of new config interface
The new datastore package is being migrated to because it fixes the previous issues of being locked to BoltDB. It would allow for other databases to be added to Mattrax easily in the future. The new configuration interface is a yaml API. This is being migrated to over GraphQL because I don't currently have time to developer a full web interface but would like to get the project in a functional state.